ファイル名変更、必要ないファイル削除
