+++ src/main/java/jp/co/aforce/dao/ManageDAO.java
@@ -2,41 +2,27 @@ package jp.co.aforce.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.util.ArrayList;
-import java.util.List;
-
-import jp.co.aforce.beans.Manage;
 
 public class ManageDAO extends DAO {
-	public List<Manage> search()
+	public int insert(String product_name, int product_price, String product_description, int product_stock)
 			throws Exception {
-		List<Manage> list = new ArrayList<>();
 
 		Connection con = getConnection();
 
 		PreparedStatement st;
 		st = con.prepareStatement(
-				"SELECT * FROM Product");
-		ResultSet rs = st.executeQuery();
+				"INSERT INTO Product (Product_Name, Product_Price, Product_Description, Product_Stock) values (?, ?, ?, ?)");
+		st.setString(1, product_name);
+		st.setInt(2, product_price);
+		st.setString(3, product_description);
+		st.setInt(4, product_stock);
+		int result = st.executeUpdate();
 
-		while (rs.next()) {
-			Manage manage = new Manage();
-			manage.setProduct_id(rs.getInt("product_id"));
-			manage.setProduct_name(rs.getString("product_name"));
-			manage.setProduct_price(rs.getInt("product_price"));
-			manage.setProduct_description(rs.getString("product_description"));
-			manage.setProduct_stock(rs.getInt("product_stock"));
-			list.add(manage);
-		}
 		st.close();
 		con.close();
 
-		return list;
-	}
+		System.out.println(result);
+		return result;
 
-	public int insert(String product_name, int product_price, String product_description, int product_stock) {
-		// TODO 自動生成されたメソッド・スタブ
-		return 0;
 	}
 }
