--- src/main/webapp/views/manageB.jsp
@@ -1,15 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"%>
-<%@include file="../header.html"%>
-
-<h1>商品管理</h1>
-
-<p>
-	<button type=button
-		onclick="location.href='/ShoppingSite/views/manage.jsp'">登録</button>
-</p>
-
-<p>
-	<button type="button" onclick="history.back()">戻る</button>
-</p>
-
-<%@include file="../footer.html"%>
